refactor: Controller 책임 분리
