+++ src/main/java/blackjack/Application.java
@@ -6,7 +6,7 @@ import blackjack.model.card.CardProvider;
 public class Application {
 
     public static void main(String[] args) {
-        BlackjackController blackjackController = new BlackjackController(new CardProvider(), new BlackjackGame());
+        BlackjackController blackjackController = new BlackjackController(new BlackjackGame(new CardProvider()));
         blackjackController.run();
     }
 }
