+++ src/main/java/blackjack/BlackjackGame.java
@@ -1,14 +1,27 @@
 package blackjack;
 
+import static blackjack.util.ExceptionHandler.retryUntilSuccess;
+
 import blackjack.model.BetAmount;
+import blackjack.model.card.CardProvider;
+import blackjack.model.card.HitCommand;
+import blackjack.model.user.Dealer;
 import blackjack.model.user.Player;
 import blackjack.model.user.Users;
+import blackjack.view.InputView;
+import blackjack.view.OutputView;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.function.Function;
 import java.util.function.Supplier;
 
 public class BlackjackGame {
+
+    private final CardProvider cardProvider;
+
+    public BlackjackGame(CardProvider cardProvider) {
+        this.cardProvider = cardProvider;
+    }
 
     public Users createUsers(Supplier<String> readUsername) {
         String input = readUsername.get();
@@ -24,5 +37,40 @@ public class BlackjackGame {
             betAmounts.add(betAmount);
         }
         return betAmounts;
+    }
+
+    public void provideInitCards(Users users) {
+        cardProvider.provideInitCards(users);
+    }
+
+    public void hit(Users users, Function<Player, String> readHitCommand) {
+        List<Player> players = users.getPlayers();
+        Dealer dealer = users.getDealer();
+
+        for (Player player : players) {
+            while (retryUntilSuccess(() -> checkY(player, readHitCommand)) && checkAddCard(player)) {
+                cardProvider.provideOneCard(player);
+                OutputView.printPlayerCards(player);
+            }
+        }
+
+        while (dealer.isHitAvailable()) {
+            cardProvider.provideOneCard(dealer);
+            OutputView.printDealerHit();
+        }
+    }
+
+    private boolean checkY(Player player, Function<Player, String> readHitCommand) {
+        String input = readHitCommand.apply(player);
+        HitCommand hitCommand = new HitCommand(input);
+        return hitCommand.isY();
+    }
+
+    private boolean checkAddCard(Player player) {
+        if (player.isHitAvailable()) {
+            return true;
+        }
+        OutputView.printCantHit();
+        return false;
     }
 }
