+++ src/main/java/blackjack/view/InputView.java
@@ -16,7 +16,7 @@ public class InputView {
         return sc.nextLine();
     }
 
-    public static String readCardAdd(Player player) {
+    public static String readHitCommand(Player player) {
         System.out.println(player.getName() + "는 한장의 카드를 더 받겠습니까?(예는 y, 아니오는 n)");
         return sc.nextLine();
     }
